add service,repository,DBConfig. change web.xml,index.jsp,servlets
