+++ src/main/java/hw2_castomArrayList/TestArray.java
@@ -1,4 +1,4 @@
-package hw1_castomArrayList;
+package hw2_castomArrayList;
 
 import java.util.Comparator;
