+++ src/main/java/hw3_smale_project/DBConstants.java
@@ -1,4 +1,4 @@
-package hw2_smale_project;
+package hw3_smale_project;
 
 public interface DBConstants {
     String DB_DRIVER = "org.postgresql.Driver";
