+++ src/main/java/hw3_smale_project/mapper/SectionMapper.java
@@ -1,12 +1,9 @@
-package hw2_smale_project.mapper;
+package hw3_smale_project.mapper;
 
-import hw2_smale_project.model.Section;
-import hw2_smale_project.model.Teacher;
+import hw3_smale_project.model.Section;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.List;
 
 public class SectionMapper {
     public Section toSection(ResultSet resultSet) throws SQLException {
