+++ src/main/java/hw3_smale_project/model/Section.java
@@ -1,18 +1,18 @@
-package hw2_smale_project.model;
+package hw3_smale_project.model;
 
-public class Teacher {
+public class Section {
+
     private int id;
-    private String name;
-    private String surname;
 
-    public Teacher() {
+    private String name;
+
+    public Section() {
 
     }
 
-    public Teacher(int id, String name, String surname) {
+    public Section(int id, String name) {
         this.id = id;
         this.name = name;
-        this.surname = surname;
     }
 
     public int getId() {
@@ -31,20 +31,11 @@ public class Teacher {
         this.name = name;
     }
 
-    public String getSurname() {
-        return surname;
-    }
-
-    public void setSurname(String surname) {
-        this.surname = surname;
-    }
-
     @Override
     public String toString() {
-        return "Teacher{" +
+        return "Section{" +
                 "id=" + id +
                 ", name='" + name + '\'' +
-                ", surname='" + surname + '\'' +
                 '}';
     }
 }
