+++ src/main/java/hw3_smale_project/service/SectionDAOImpl.java
@@ -1,0 +1,29 @@
+package hw3_smale_project.service;
+
+import hw3_smale_project.model.Section;
+import hw3_smale_project.repository.SectionDAO;
+
+import java.sql.SQLException;
+
+public class SectionDAOImpl {
+
+    private final SectionDAO sectionDAO;
+
+
+    public SectionDAOImpl(SectionDAO sectionDAO) {
+        this.sectionDAO = sectionDAO;
+    }
+
+
+    public Section findSectionByTeacherId(int id) {
+        Section section = new Section();
+
+        try {
+            section = sectionDAO.findSectionByTeacherId(id);
+        } catch (SQLException | ClassNotFoundException e) {
+            throw new RuntimeException(e);
+        }
+        System.out.println(section);
+        return section;
+    }
+}
